# libs/go/apidoc/openapi.yaml
components:
  securitySchemes:
    elvish_session:
      in: cookie
      name: elvish_session
      type: apiKey
info:
  description: JSON API under /api. Regenerate this file with `make openapi` (see docs/openapi/supplemental.yaml for routes not expressed as literal `p ==` cases in api.go).
  title: ELVish HTTP API
  version: 0.0.0
openapi: 3.0.3
paths:
...
  /api/admin/users/{userId}:
    delete:
      operationId: delete_api_admin_users_userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Delete user (admin; full purge)
    get:
      operationId: get_api_admin_users_userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Get user (admin)
    patch:
      operationId: patch_api_admin_users_userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Patch user (admin; is_admin)
  /api/admin/users/{userId}/disable:
    post:
      operationId: post_api_admin_users_userId_disable
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Disable user (admin)
  /api/admin/users/{userId}/enable:
    post:
      operationId: post_api_admin_users_userId_enable
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Re-enable user (admin; SRP accounts only)
...
  /api/v1/filters/{filterId}:
    delete:
      operationId: delete_api_v1_filters_filterId
      parameters:
        - in: path
          name: filterId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Delete mail filter
    put:
      operationId: put_api_v1_filters_filterId
      parameters:
        - in: path
          name: filterId
          required: true
          schema:
            type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
          description: Response (JSON or error body; see handler and docs/e2ee-mail-spec.md for mail).
      summary: Update mail filter
...
servers:
  - url: /